Move aiBridge and secretstore out of engine/ folder
aiBridge.{h,cpp}: Desktop/engine/ → Desktop/ai/
secretstore.{h,cpp}: Desktop/engine/ → Desktop/utilities/

Neither class is related to the engine subsystem. AiBridge is
an AI chat bridge and secretstore is a general utility for
API key storage. Engine/ now contains only EngineSync,
EngineRepresentation, and RScriptStore.
