+++ Desktop/gui/preferencesmodel.cpp
@@ -1,5 +1,5 @@
 #include "preferencesmodel.h"
-#include "engine/secretstore.h"
+#include "utilities/secretstore.h"
 #include "utilities/qutils.h"
 #include "utilities/settings.h"
 #include "utilities/messageforwarder.h"
